auto-generate action_groups
