+++ docs/requirements.txt
@@ -3,3 +3,4 @@ ansible-base>=2.10.0rc2
 sphinx-rtd-theme
 git+https://github.com/felixfontein/ansible-basic-sphinx-ext
 recommonmark
+py
