+++ requirements-lint.txt
@@ -1,3 +1,4 @@
+py
 flake8<4
 yamllint
 ansible-lint
