+++ tests/test_module_state.py
@@ -1,19 +1,7 @@
-import py.path
 import pytest
 
 from .conftest import TEST_PLAYBOOKS
-
-MODULES_PATH = py.path.local(__file__).realpath() / '..' / '..' / 'plugins' / 'modules'
-
-
-def find_all_modules():
-    for module in MODULES_PATH.listdir(sort=True):
-        module = module.basename
-        if module.endswith('.py') and not module.startswith('_'):
-            yield module.replace('.py', '')
-
-
-ALL_MODULES = list(find_all_modules())
+from .utils import ALL_MODULES, MODULES_PATH
 
 
 def _module_file_path(module):
